fix bug on raspi
After various debugging I finally found that the keys returned from
getkey() was different than the one for addkey(). As a consequence,
the decryption and encryption were done with different keys.
Stupid mistake, the memory allocated for an entry in the keymap was
equal to a pointer and not to the whole struct. What is very surprising
though is that it only led to out of bounds reading on the raspi and not
on my laptop.

If I used getkey() after addkey for encryption I would probably never have
found this bug before testing with multiple clients.